requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
Brotli==1.1.0
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==41.0.7
cssselect2==0.7.0
daphne==4.0.0
Django==4.2.7
django-admin-argon-dashboard==1.0.16
django-weasyprint==2.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
fcm-django==2.0.0
firebase-admin==6.3.0
fonttools==4.47.2
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
html5lib==1.1
httplib2==0.22.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
itsdangerous==0.24
msgpack==1.0.7
Pillow==10.1.0
proto-plus==1.23.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydyf==0.8.0
PyJWT==2.8.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
pyphen==0.14.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
rsa==4.9
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
stripe==7.11.0
tinycss2==1.2.1
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
weasyprint==60.1
webencodings==0.5.1
webservices==0.7
whitenoise==6.6.0
zope.interface==6.1
zopfli==0.2.3
